Migrate crosstests to use testcontainers
Instead of manually having to exclude tests from `./gradlew build`
(since the crosstest tests fail if the Docker server isn't running),
migrate the crosstests to use testcontainers to spin up the server so
the tests can always run. This also simplifies the makefile - we no
longer need to manually start/stop the crosstests server and also don't
need to bind to static ports on the host - it will use ephemeral ports.