Improve code generation in connect-kotlin
Generate code to the build/generated/sources directory instead of under
src/ to avoid the need for manual cleanup and consolidate where code is
placed (so it can be ignored from spotless checks).

Remove generated code from googleapis from source control - we were
already generating it with buf.gen.yaml but never configuring the build
to use the generated code.