Rename and restructure grpcCompletionToConnectError
In most (3 of 4) uses of this method, the first argument
was known to be non-null at the call site. Since in the
event that the first argument _was_ null, we basically create
a `ConnectException` without much helpful information about
any error, it made sense to me to remove the concept of a nullable
completion from the method entirely.

The function is also a little more idiomatic now, following
the typical `to<Type>OrNull` pattern established by the language for
conversion between types.

It also reduces the classes generated by the compiler since
the method is a proper member of `GRPCCompletion`.

I would probably push this idea a little further in how
error states are modeled more generally, but that's creeping
outside the scope of this change.