Allow building with Java 20 and later versions (#146)
Update the KotlinCompile args to add `-Xjdk-release=1.8` which is the
equivalent of the `--release 8` argument when compiling Java code. This
will ensure that we only use APIs which are in Java 1.8 and not added in
later versions. This fixes compiler warnings around use of `new
URL(String)` which is deprecated starting in Java 20.