Add some additional E2E tests for client
In several cases, we were not correctly calling callbacks when an
exception occurred while reading a response body, parsing it using a
Codec, or attempting to decompress it. Add some additional E2E tests in
the new MockWebServerTests to exercise these failures and update the
library code to handle the exceptions appropriately.