Add method stream type to generated code (#172)
It is useful to have the stream type in the generated code (similar to
other Connect implementations). In the future we may use this
information to optimize stream handling - for example, avoiding the
pipe/duplex/oneshot for server streaming calls to enable Connect over
HTTP/1.1. Update the generator to include the stream type in the
generated code.

Additionally, move the HTTP method to the HTTPRequest class - mutating a
method spec to determine which HTTP method to use seems wrong. We should
keep MethodSpec to contain only fields set from the generator.