Adds new conformance/client module (#182)
This PR is purely additive. It does not make any changes to the existing
container/crosstest-based conformance test targets. It's just starting
to add code to eventually replace all of that.

So the old conformance tests are in these modules:
- `conformance/common`
- `conformance/google-java`
- `conformance/google-javalite`

The new ones are in these:
- `conformance/client`
- `conformance/client/google-java`
- `conformance/client/google-javalite`

This doesn't include any actual logic yet. This was a rather substantial
PR already, so I figured it would be good to get this approach reviewed
before fleshing out the conformance client handler logic (which are all
stubbed out in `Client.kt` with TODOs).

To avoid the redundant implementations of the current conformance tests,
I'm doing the following:
1. Generate all conformance protos, using lite runtime, into the
`conformance/client` module. So the shared logic deals with the lite
generated message types, since that's the lowest common denominator.
(The full runtime should be able to correctly support lite generate
code.)
2. This means that the `conformance/client/google-javalite` module
doesn't need any add'l generated code.
3. The `conformance/client/google-java` module, on the other hand, does.
So the shared code will convert from generated lite messages to
normal/non-lite messages. This uses a `buf.gen.lite.yaml` which has a
different package prefix, so both sets of generated messages can
co-exist. The dependencies of the non-lite module exclude the javalite
protobuf runtime, swapping in the full runtime instead.