Some cleanup, mostly in the HTTP request representations (#211)
1. The biggest chunk clarifies the representation of HTTP requests.
Previously the `HTTPRequest` class _might_ have a request message.
This message was really required for unary operations (and, if absent,
would be treated as empty request) and ignored for stream operations.
So now the type is split into two: a base`HTTPRequest` which has no
request message, and a `UnaryHTTPRequest` which has a non-optional
message type. Also, since everything in the framework works with
`Buffer` for messages, this changes the type of the message from
`ByteArray` to `Buffer`.
2. The next change renames some variables/parameters in the
compression stuff to make it a little easier to read.
3. Another small change tries to make `Envelope.pack` a little more
DRY.
4. The final change is a bug fix: when a gRPC operation completes,
this was treating missing trailers as a successful RPC, as if the trailers
were present and indicated a status of "ok". But that is not correct as
missing trailers in the gRPC protocol means something is definitely
wrong (even if it's a unary operation that includes the singular
response message). So now the client will consider this case to
be an RPC error.