Provide way to auto-close streams (#213)
This makes the three stream types closeable. They
aren't `java.io.Closeable` because I want the close to be suspending
function. (Not currently necessary, but it will be useful for
something I want to do coming soon.)

This also changes the main method signatures to invoke the
RPC and use the stream in a block, so that the stream is automatically
closed at the end.

This is still all just in the conformance client. Once we're happy with
the shape of the APIs there, I'd like to re-do the actual interfaces in
the main com.connectrpc package. So I consider the internal APIs
of the conformance adapter package an experimental playground.