Auto-dispatch to I/O context (#218)
This adds a new property to `ProtocolClientConfig` called
`ioCoroutineContext`. It defaults to null, which means there is no
change in behavior from the current defaults -- operations happen in the
calling coroutine context, so callers of RPC methods must handle the
dispatch.

But when it is configured as non-null (most likely using a value like
`Dispatchers.IO`), then I/O and blocking operations for an RPC are
automatically dispatched to the given coroutine context. This allows for
RPC client code to be used from any coroutine context without having to
worry about explicitly dispatching.