Denormalize "Code", slight simplification of HTTP interface (#224)
1. This denormalizes the presence of code. It was previously present
both on `StreamResult.Complete` and `ResponseMessage`, even though it
was redundant with the code on any non-null `ConnectException`.
2. Removes `Code.OK`. It is now just an error code. An "OK" result is
easily distinguished by the lack of an exception.
3. The `HTTPClientInterface` was previously responsible for translating
HTTP status -> `Code` and the `HTTPResponse` reported back the status in
an optional "trace info" field. But translating status to `Code` should
be the responsibility of the protocol implementations, not the HTTP
client. (Though it is still up to the HTTP client to classify exceptions
that prevent an HTTP response from being received, as these could be
client-implementation-specific exception types). So this change promotes
the status code out of trace info and into the `HTTPResponse`.
4. `ProtocolClient.stream` used `suspendCancellableCoroutine`
incorrectly. It looks like it was trying to associate the stream with
the coroutine, so if the coroutine is cancelled then the stream is
closed. But it is only effective *while that method is running*, which
is almost none at all since it immediately returns a result (no
suspensions).