Simplify UnaryBlockingCall (#225)
* It's now just an interface, and the implementation is in the impl
package (just like the stream interfaces).
* The implementation is simpler and thread-safe. No more mutator methods
needed.
* It no longer supports calling it more than once (which it _ostensibly_
did before, according to doc comments). This was incorrectly implemented
anyway since, once invoked a subsequent time, it became impossible to
cancel earlier invocations (and was not thread-safe). It is much easier
to disallow multiple invocations than to try to make it properly support
them. It's also much easier to reason about IMO.