Fix all conformance failures other than timeouts/deadlines (#274)
* The first fix is for how trailers-only responses are classified.
This was previously just looking for a "grpc-status" key in the headers.
If it was present, it was treating it as a trailers-only response, even
if there was a body and/or trailers.
* The second fix is so that the client reports errors in the face of
unexpected response content types. With a little code reorganization,
we can improve this logic in the future, to increase code sharing
(especially between gRPC and gRPC-Web).